Renaming functions for errors in UE 5.1
Originally tested in 5.0 without issue, but in 5.1 I encountered errors for MoveFile, CopyFile, DeleteFile and SendRequest. Example of an error that was encountered would say: "C2039 'CopyFileW': is not a member of BrainCloudGroupFile". I've seen the behavior from UE before where they insert a random letter if theres already a function named the same, so the fix here is to have all function start lowercase so that there's no conflict.