Bug fix for RTT
Whenever I tested with Relay Test App, I would get an engine crash after a game session has ended. I tracked down the reason for the crash was because the plugin did not clean up registered callbacks so when it tries to execute callback, the engine crashes. This fix ensures when websocket executes onClose() to clean up callbacks if it didn't get disconnected for another reason (IE: Bad connection).