Fixing bugs with failing requests right away for bad data.
Found bug when running all unit tests and Push Notification kept getting hanged up. After investigating the problem, it seems that we cannot execute async functions in this manner and it makes UE hang on the one test until it times out. So the solution here is to wait a frame to then execute the callback so by that time the callback is on the stack when its executed.